split SingleResponseView from inbox/index
